feat: Enable HTTPS and mount certs in Docker
This commit enables HTTPS for the application by setting the HTTPS_ENABLED environment variable to true in the docker-compose.yml file.

It also mounts the local ./certs directory into the container at /app/certs, allowing the application to access the necessary SSL certificates. This is a necessary step to run the application with HTTPS in a development environment.